历史记录界面更新，优化代码

+++ Settings/ResetItem.xaml.cs
@@ -1,5 +1,6 @@
 using Edge.Data;
 using Microsoft.UI.Xaml.Controls;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
@@ -14,12 +15,16 @@ namespace Edge
             this.InitializeComponent();
         }
 
-        private void ResetUserSettings(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
+        private async void ResetUserSettings(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
         {
-            string path = ApplicationData.Current.LocalFolder.Path + "/settings.json";
-            File.Delete(path);
-            Info.CheckUserSettingData();
-            App.settings = JsonSerializer.Deserialize<Dictionary<string, JsonElement>>(File.ReadAllText(path));
+            ContentDialogResult res = await resetDialog.ShowAsync();
+            if (res == ContentDialogResult.Primary)
+            {
+                string path = ApplicationData.Current.LocalFolder.Path + "/settings.json";
+                File.Delete(path);
+                Info.CheckUserSettingData();
+                App.settings = JsonSerializer.Deserialize<Dictionary<string, JsonElement>>(File.ReadAllText(path));
+            }
         }
     }
 }
